requirements.txt
torch==2.3.1
transformers==4.44.2
sentence-transformers==3.0.1
huggingface-hub==0.24.6
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.5
textblob==0.18.0.post0
nltk==3.8.1
spacy==3.7.4
openai-whisper==20231117
ffmpeg-python==0.2.0
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.2.post1
PyPDF2==3.0.1
pdfplumber==0.10.4
streamlit==1.39.0
python-dotenv==1.0.1
language-tool-python==2.7.1
reportlab==4.2.0
requests==2.32.3
